# backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: EduPilot Gateway API
  version: 0.14.0
  description: |
    Hợp đồng API của gateway Go — NGUỒN SỰ THẬT (D45). Mỗi route trong mã phải có ở đây và ngược lại
    (`internal/contract` kiểm hai chiều). Route thử nằm ở `openapi.test.yaml`, chỉ có trong binary dựng bằng
    build tag `testroutes`. Phase PG chưa có endpoint nghiệp vụ; P1 (FEAT-llm-gateway US-P1-04) thêm 13 thao tác `/admin/llm/*` và `/courses/{id}/llm-budget`.
    P2 (FEAT-account-security US-P2-02) thêm 9 thao tác `/auth/{login,refresh,logout,register,verify-email,resend-verification,forgot-password,reset-password,tokens/preview}` và 4 thao tác `/me/{password,sessions,sessions/{id}}` (US-P2-04); US-P2-05 thêm 429 (`LOGIN_THROTTLED` | `RATE_LIMITED`) cho login / register / verify-email / reset-password / refresh: công khai (không Bearer); `refresh`/`logout` dùng cookie `ep_rt` (HttpOnly) + kiểm Origin.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/events:
    get:
      operationId: streamEvents
      tags: [events]
      summary: Luồng sự kiện thời gian thực (SSE) của chính người gọi
      description: |
        `text/event-stream`. Xác thực bằng header `Authorization` (không nhận token qua query). Khung sự kiện
        `id/event/data`; heartbeat `: hb` mỗi 25 s; tối đa 2 kết nối mỗi người; tự đóng sau 120 s bằng
        `event: reconnect`. Nối lại bằng header `Last-Event-ID`.
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Luồng SSE đang mở
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/me/exam-lock:
    get:
      operationId: getMyExamLock
      tags: [me]
      summary: "Chat AI có đang bị khoá vì lượt làm bài không (chỉ chính mình; không lộ tên bài thi / lớp)"
      responses:
        "200":
          description: Trạng thái khoá
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamLock"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /api/v1/courses/{id}/exams/{eid}:
    get:
      operationId: getExam
      tags: [exam]
      summary: "Chi tiết bài thi (Staff: đầy đủ kèm mục; Sinh viên: bản giới hạn, DRAFT → 404)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Bài thi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetailOrStudent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateExam
      tags: [exam]
      summary: "Sửa bài thi (khoá lạc quan `version`; ngoài DRAFT chỉ title / instructions / reveal_answers / appeal_days)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "200":
          description: Bài sau khi sửa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    delete:
      operationId: deleteExam
      tags: [exam]
      summary: "Xoá bài thi nháp (Giảng viên; cascade mục)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Đã xoá
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/courses/{id}/exams/{eid}/attempts/{aid}/events:
    post:
      operationId: postExamEvents
      tags: [exam]
      summary: "Gửi sự kiện rời tab / dán (không nội dung; lô tối đa 50, phần dư bị bỏ; không bao giờ làm hỏng việc lưu bài)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: aid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamEventsRequest"
      responses:
        "204":
          description: Đã nhận (hoặc bỏ qua khi lượt không còn làm dở)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/exams/{eid}/events:
    get:
      operationId: listExamEvents
      tags: [exam]
      summary: "Log rời tab / dán của một lượt — CHỈ Giảng viên (TA, sinh viên kể cả của chính mình, Admin → 403)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attempt
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Tóm tắt + sự kiện (mới nhất trước)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamEventsPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/exams/{eid}/similarity:
    get:
      operationId: listSimilarity
      tags: [exam]
      summary: "Cặp bài code nghi giống nhau (điểm cao trước; mặc định bản chạy mới nhất) — chỉ Giảng viên; tín hiệu để tham khảo"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: run
          in: query
          schema:
            type: string
            format: uuid
        - name: flagged
          in: query
          schema:
            type: boolean
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Danh sách
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarityPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/exams/{eid}/similarity/run:
    post:
      operationId: runSimilarity
      tags: [exam]
      summary: "Chạy lại so độ giống (bài đã đóng; việc nền)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Đã xếp việc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/exams/{eid}/similarity/{sid}:
    get:
      operationId: getSimilarity
      tags: [exam]
      summary: "Một cặp kèm hai mã cạnh nhau và các dòng khớp"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Cặp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarityDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/courses/{id}/exams/{eid}/similarity/{sid}/review:
    put:
      operationId: reviewSimilarity
      tags: [exam]
      summary: "Đánh dấu Đã xem — không có vấn đề (CLEARED) hoặc Cần trao đổi (FOLLOW_UP); không có hành động trừ điểm"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimilarityReviewRequest"
      responses:
        "200":
          description: Đã đánh dấu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarityView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Chưa đăng nhập, token hết hạn / không hợp lệ / phiên bị thu hồi (`UNAUTHENTICATED`, `TOKEN_EXPIRED`, `TOKEN_INVALID`, `SESSION_REVOKED`, `INVALID_CREDENTIALS`)
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Sai vai trò, ngoài lớp hoặc Origin lạ (`FORBIDDEN`, `details.reason` = `role` | `course` | `origin`); tài khoản bị khoá (`ACCOUNT_DISABLED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Không có route / tài nguyên / không có quyền biết (`NOT_FOUND`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Xung đột trạng thái hoặc phiên bản (`CONFLICT`, `VERSION_CONFLICT`, `IDEMPOTENCY_IN_PROGRESS`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Dữ liệu không hợp lệ (`VALIDATION_FAILED`, `INVALID_CURSOR`, `IDEMPOTENCY_KEY_*`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Quá hạn mức (`RATE_LIMITED`, `SSE_LIMIT_REACHED`); kèm `Retry-After`
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: Phụ thuộc hỏng (`SERVICE_UNAVAILABLE`, `NOT_READY`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message, trace_id]
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
            - UNAUTHENTICATED
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VERSION_CONFLICT
            - IDEMPOTENCY_IN_PROGRESS
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - VALIDATION_FAILED
            - INVALID_CURSOR
            - IDEMPOTENCY_KEY_REUSED
            - IDEMPOTENCY_KEY_REQUIRED
            - RATE_LIMITED
            - SSE_LIMIT_REACHED
            - INTERNAL
            - SERVICE_UNAVAILABLE
            - NOT_READY
            - DEADLINE_EXCEEDED
            - OVERLOADED
            - LLM_NOT_CONFIGURED
            - LLM_UNAVAILABLE
            - PROVIDER_IN_USE
            - MODEL_DIMS_MISMATCH
            - ROUTE_INVALID
            - INVALID_CREDENTIALS
            - LOGIN_THROTTLED
            - ACCOUNT_DISABLED
            - SESSION_REVOKED
            - LINK_INVALID
            - EMAIL_NOT_VERIFIED
            - JOIN_CODE_INVALID
            - COURSE_FULL
            - COURSE_ARCHIVED
            - EXAM_NOT_OPEN
            - ATTEMPT_ALREADY_SUBMITTED
            - ATTEMPT_CLOSED
            - ATTEMPT_OTHER_TAB
            - DRAFT_CONFLICT
            - EXAM_LOCKED
            - QUESTION_IN_USE
            - RESULT_NOT_PUBLISHED
            - SUBMISSION_LIMIT_REACHED
            - JUDGE_UNAVAILABLE
            - APPEAL_WINDOW_CLOSED
            - APPEAL_EXISTS
        message:
          type: string
        trace_id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        details: {}
        retry_after:
          type: integer
          minimum: 1
    JobAccepted:
      type: object
      required: [job_id]
      additionalProperties: false
      properties:
        job_id:
          type: string
          format: uuid
...
    ExamInput:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        instructions:
          type: string
        opens_at:
          type: string
          format: date-time
        closes_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        shuffle_questions:
          type: boolean
        shuffle_options:
          type: boolean
        max_score:
          type: string
        rounding_step:
          type: string
        multi_scoring:
          type: string
          enum: [PARTIAL, ALL_OR_NOTHING]
        reveal_answers:
          type: boolean
        appeal_days:
          type: integer
        version:
          type: integer
...
    ExamAttempts:
      type: object
      required: [started, graded]
      additionalProperties: false
      properties:
        started:
          type: integer
        graded:
          type: integer
...
    ExamItem:
      type: object
      required: [id, question_id, position, points, type, title, topic, difficulty, review_status, archived]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        position:
          type: integer
        points:
          type: string
        type:
          type: string
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
        review_status:
          type: string
        archived:
          type: boolean
    ExamDetail:
      type: object
      required: [id, title, kind, status, effective_status, opens_at, closes_at, duration_minutes, items_count, attempts, published_at, version, created_at,
                 instructions, shuffle_questions, shuffle_options, max_score, rounding_step, multi_scoring, reveal_answers, appeal_days, publish_hold, regrading, items, created_by, updated_at]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        kind:
          type: string
          enum: [MCQ, CODE, MIXED]
        status:
          type: string
          enum: [DRAFT, SCHEDULED, OPEN, CLOSED, PUBLISHED]
        effective_status:
          type: string
          enum: [DRAFT, SCHEDULED, OPEN, CLOSED, PUBLISHED]
        opens_at:
          type: [string, "null"]
          format: date-time
        closes_at:
          type: [string, "null"]
          format: date-time
        duration_minutes:
          type: [integer, "null"]
        items_count:
          type: integer
        attempts:
          $ref: "#/components/schemas/ExamAttempts"
        published_at:
          type: [string, "null"]
          format: date-time
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        instructions:
          type: [string, "null"]
        shuffle_questions:
          type: boolean
        shuffle_options:
          type: boolean
        max_score:
          type: string
        rounding_step:
          type: string
        multi_scoring:
          type: string
        reveal_answers:
          type: boolean
        appeal_days:
          type: integer
        publish_hold:
          type: boolean
        regrading:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExamItem"
        created_by:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
    MyAttempt:
      type: object
      required: [id, status, deadline_at, submitted_at]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        deadline_at:
          type: string
          format: date-time
        submitted_at:
          type: [string, "null"]
          format: date-time
    ExamStudentView:
      type: object
      required: [id, title, instructions, kind, opens_at, closes_at, duration_minutes, max_score, status, my_attempt, my_score]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        instructions:
          type: [string, "null"]
        kind:
          type: string
          enum: [MCQ, CODE, MIXED]
        opens_at:
          type: [string, "null"]
          format: date-time
        closes_at:
          type: [string, "null"]
          format: date-time
        duration_minutes:
          type: [integer, "null"]
        max_score:
          type: string
        status:
          type: string
          enum: [DRAFT, SCHEDULED, OPEN, CLOSED, PUBLISHED]
        my_attempt:
          oneOf:
            - $ref: "#/components/schemas/MyAttempt"
            - type: "null"
        my_score:
          type: [string, "null"]
    ExamDetailOrStudent:
      oneOf:
        - $ref: "#/components/schemas/ExamDetail"
        - $ref: "#/components/schemas/ExamStudentView"
...
    ExamEventsRequest:
      type: object
      required: [events]
      additionalProperties: false
      properties:
        events:
          type: array
          items:
            type: object
            required: [type]
            additionalProperties: false
            properties:
              type:
                type: string
                description: "TAB_HIDDEN | TAB_VISIBLE | PASTE | OFFLINE | ONLINE (loại khác bị bỏ)"
              client_at:
                type: [string, "null"]
                format: date-time
              meta:
                type: object
                description: "Chỉ giữ duration_ms, chars (số nguyên), item_id (uuid); khoá khác bị bỏ; tối đa 300 byte"
    ExamIntegritySummary:
      type: object
      required: [tab_hidden_count, tab_hidden_ms, paste_count, paste_chars, offline_count, takeover_count, chat_blocked_count]
      additionalProperties: false
      properties:
        tab_hidden_count:
          type: integer
        tab_hidden_ms:
          type: integer
        paste_count:
          type: integer
        paste_chars:
          type: integer
        offline_count:
          type: integer
        takeover_count:
          type: integer
        chat_blocked_count:
          type: integer
    ExamEventView:
      type: object
      required: [id, type, occurred_at, client_at, meta]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [TAB_HIDDEN, TAB_VISIBLE, PASTE, OFFLINE, ONLINE, TAB_TAKEOVER, CHAT_BLOCKED]
        occurred_at:
          type: string
          format: date-time
        client_at:
          type: [string, "null"]
          format: date-time
        meta:
          type: object
    ExamEventsPage:
      type: object
      required: [summary, items, next_cursor]
      additionalProperties: false
      properties:
        summary:
          $ref: "#/components/schemas/ExamIntegritySummary"
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/ExamEventView"
        next_cursor:
          type: [string, "null"]
    ExamLock:
      type: object
      required: [locked]
      additionalProperties: false
      properties:
        locked:
          type: boolean
        until:
          type: string
          format: date-time
    SimilarityStudent:
      type: object
      required: [attempt_id, name]
      additionalProperties: false
      properties:
        attempt_id:
          type: string
          format: uuid
        name:
          type: string
    SimilarityView:
      type: object
      required: [id, problem_id, problem_title, run_id, a, b, score, shared_fingerprints, flagged, review_state, note, reviewed_at, created_at]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        problem_id:
          type: string
          format: uuid
        problem_title:
          type: string
        run_id:
          type: string
          format: uuid
        a:
          $ref: "#/components/schemas/SimilarityStudent"
        b:
          $ref: "#/components/schemas/SimilarityStudent"
        score:
          type: string
          description: "Jaccard dạng thập phân 0.000–1.000"
        shared_fingerprints:
          type: integer
        flagged:
          type: boolean
        review_state:
          type: string
          enum: [NEW, CLEARED, FOLLOW_UP]
        note:
          type: [string, "null"]
        reviewed_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
    SimilarityPage:
      type: object
      required: [items, next_cursor]
      additionalProperties: false
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/SimilarityView"
        next_cursor:
          type: [string, "null"]
    SimilaritySide:
      type: object
      required: [language, source, match_lines]
      additionalProperties: false
      properties:
        language:
          type: string
        source:
          type: string
        match_lines:
          type: array
          items:
            type: integer
    SimilarityDetail:
      type: object
      required: [pair, a, b]
      additionalProperties: false
      properties:
        pair:
          $ref: "#/components/schemas/SimilarityView"
        a:
          $ref: "#/components/schemas/SimilaritySide"
        b:
          $ref: "#/components/schemas/SimilaritySide"
    SimilarityReviewRequest:
      type: object
      required: [state]
      additionalProperties: false
      properties:
        state:
          type: string
          enum: [CLEARED, FOLLOW_UP]
        note:
          type: [string, "null"]
          maxLength: 500
...